Feature/new portfolio

+++ routes/web.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Http\Controllers\{ AdminController, BadgesController, BlogController, DashboardController };
+use App\Http\Controllers\{ AdminController, BadgesController, BlogController, DashboardController, TagController};
 use App\Http\Controllers\Admin\BlogController as AdminBlogController;
 use App\Http\Controllers\Admin\DashboardController as AdminDashboardController;
 use App\Http\Controllers\Admin\PersonalController;
@@ -45,11 +45,11 @@ Route::get('/home', function () {
     return redirect('/admin');
 });
 
-Route::post('/admin/personal/{personal}', [PersonalController::class, 'update']);
+Route::post('/admin/personal/{personal}', [PersonalController::class, 'update'])->middleware('auth');
 
-Route::post('/admin/blogs/setStatus', [AdminBlogController::class, 'setStatus']);
-Route::get('/admin/blogs/paginate', [AdminBlogController::class, 'paginate']);
-Route::resource('/admin/blogs', AdminBlogController::class);
+Route::post('/admin/blogs/setStatus', [AdminBlogController::class, 'setStatus'])->middleware('auth');
+Route::get('/admin/blogs/paginate', [AdminBlogController::class, 'paginate'])->middleware('auth');
+Route::resource('/admin/blogs', AdminBlogController::class)->middleware('auth');
 
 Route::controller(AuthController::class)->group(function () {
     Route::get('auth/github', 'redirectToProvider');
@@ -57,10 +57,16 @@ Route::controller(AuthController::class)->group(function () {
 });
 
 Route::controller(AdminController::class)->group(function () {
-    Route::get('/admin/create', 'create')->name('adminCreate');
-    Route::post('/admin/store', 'store')->name('adminStore');
-    Route::post('/admin/setState', 'setState')->name('adminSetState');
-    Route::put('/admin/{id}', 'update')->name('adminUpdate');
-    Route::delete('/admin/{id}', 'delete')->name('adminDelete');
-    Route::get('/admin/{id}/edit', 'edit')->name('adminEdit');
+    Route::get('/admin/create', 'create')->name('adminCreate')->middleware('auth');
+    Route::post('/admin/store', 'store')->name('adminStore')->middleware('auth');
+    Route::post('/admin/setState', 'setState')->name('adminSetState')->middleware('auth');
+    Route::put('/admin/{id}', 'update')->name('adminUpdate')->middleware('auth');
+    Route::delete('/admin/{id}', 'delete')->name('adminDelete')->middleware('auth');
+    Route::get('/admin/{id}/edit', 'edit')->name('adminEdit')->middleware('auth');
+});
+
+Route::controller(TagController::class)->group(function () {
+    Route::get('/admin/tag', 'index')->name('tagIndex')->middleware('auth');
+    Route::post('/admin/tag', 'store')->name('tagStore')->middleware('auth');
+    Route::delete('/admin/tag/{tag}', 'delete')->name('tagDelete')->middleware('auth');
 });
